requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-filter==2.4.0
django-froala-editor==3.2.2
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
Markdown==3.3.4
newsapi-python==0.2.6
Pillow==8.3.1
platformdirs==2.3.0
psycopg2==2.9.1
pytz==2021.1
qrcode==7.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.7.2
whitenoise==5.3.0
